Minor change of ALL_SUITS from set to tuple
There are some places where we assume that a fresh deck is already
sorted (mostly tests). This will ensure that always, where before
the (unordered) set might have occasionally not have iterated the same
way between executions. There would not have been any issues from this
right now, because within an execution it would have always resolved the
same way, and that's enough for what the code needed. But this will be
better long term, and the cost of occasionally iterating the tuple of 4
elements is not meaningful enough to need the constant-time lookup of
the set.